QA-12026: defined a new ObjectMapper to read the template-engine.yml property and injected in the proper strategy
