+++ pom.xml
@@ -64,6 +64,11 @@
     </properties>
 
     <dependencies>
+        <dependency>
+            <groupId>com.fasterxml.jackson.dataformat</groupId>
+            <artifactId>jackson-dataformat-yaml</artifactId>
+            <version>2.11.4</version>
+        </dependency>
 
         <dependency>
             <groupId>it.pagopa.pn</groupId>
