+++ src/test/java/it/pagopa/pn/cucumber/steps/CucumberSpringIntegration.java
@@ -1,8 +1,8 @@
 package it.pagopa.pn.cucumber.steps;
 
 import io.cucumber.spring.CucumberContextConfiguration;
-import it.pagopa.pn.client.b2b.pa.config.PnB2bClientTemplateEngineConfigs;
 import it.pagopa.pn.client.b2b.pa.config.PnB2bClientTimingConfigs;
+import it.pagopa.pn.client.b2b.pa.config.TemplateEngineConfigBean;
 import it.pagopa.pn.client.b2b.pa.config.springconfig.*;
 import it.pagopa.pn.client.b2b.pa.parsing.config.PnLegalFactTokenProperty;
 import it.pagopa.pn.client.b2b.pa.parsing.config.PnLegalFactTokens;
@@ -182,8 +182,8 @@ import org.springframework.scheduling.annotation.EnableScheduling;
         EventTimelineParser.class,
         PnMandateAppIoClientImpl.class,
         ReworkTimelineClientImpl.class,
-        PnB2bClientTemplateEngineConfigs.class,
         LambdaInvoker.class,
+        TemplateEngineConfigBean.class,
 })
 @EnableScheduling
 @EnableConfigurationProperties
